http_server: split session handlers into submodules
