+++ daemon/http_server/src/handlers/sessions/create.rs
@@ -1,0 +1,62 @@
+use crate::AppState;
+use axum::{Json, extract::State, http::StatusCode};
+use serde::Deserialize;
+
+#[derive(Deserialize, Default)]
+pub struct CreateRequest {
+    #[serde(default)]
+    name: Option<String>,
+}
+
+pub async fn create(
+    State(state): State<AppState>,
+    Json(body): Json<CreateRequest>,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let id = state.create_session(body.name).await;
+    (StatusCode::CREATED, Json(serde_json::json!({ "id": id })))
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::test_helpers::{fresh_state, router_with};
+    use axum::body::{Body, to_bytes};
+    use axum::http::{Request, StatusCode};
+    use tower::ServiceExt;
+
+    #[tokio::test]
+    async fn create_returns_201_with_id() {
+        let (router, _) = router_with(fresh_state());
+        let resp = router
+            .oneshot(
+                Request::builder()
+                    .method("POST")
+                    .uri("/sessions")
+                    .header("content-type", "application/json")
+                    .body(Body::from(r#"{"name":"test"}"#))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(resp.status(), StatusCode::CREATED);
+        let body = to_bytes(resp.into_body(), usize::MAX).await.unwrap();
+        let v: serde_json::Value = serde_json::from_slice(&body).unwrap();
+        assert!(v["id"].is_string());
+    }
+
+    #[tokio::test]
+    async fn create_without_name_still_returns_201_with_id() {
+        let (router, _) = router_with(fresh_state());
+        let resp = router
+            .oneshot(
+                Request::builder()
+                    .method("POST")
+                    .uri("/sessions")
+                    .header("content-type", "application/json")
+                    .body(Body::from("{}"))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(resp.status(), StatusCode::CREATED);
+    }
+}
